+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Makefile clean target also removes Gemfile.lock file
+
+### Fixed
+- Fixed outdated Gemfile.lock file
+
 ## 1.2.0 - 2019-10-16
 ### Added
 - Add InSpec test for checking if AEM repository migration was skipped [#6]